Integrate deserializer_state.{c,h} into other source files
(into deserializer_callbacks.{c,h} and deserializer.{c,h})
